backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.2
pydantic-settings==2.8.1
SQLAlchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.2.0

